feat(cli): change teamsfx manifest update to teamsfx deploy manifest
