+++ packages/cli/src/cmds/deploy.ts
@@ -15,6 +15,7 @@ import {
   Inputs,
   MultiSelectQuestion,
   OptionItem,
+  Func,
 } from "@microsoft/teamsfx-api";
 
 import activate from "../activate";
@@ -36,6 +37,7 @@ export default class Deploy extends YargsCommand {
   public readonly description = "Deploy the current application.";
 
   public readonly deployPluginNodeName = constants.deployPluginNodeName;
+  public readonly subCommands: YargsCommand[] = [new DeployManifest()];
 
   public builder(yargs: Argv): Argv<any> {
     this.params = HelpParamGenerator.getYargsParamForHelp(Stage.deploy);
@@ -51,6 +53,11 @@ export default class Deploy extends YargsCommand {
         yargs.options(name, this.params[name]);
       }
     }
+
+    this.subCommands.forEach((cmd) => {
+      yargs.command(cmd.command, cmd.description, cmd.builder.bind(cmd), cmd.handler.bind(cmd));
+    });
+
     return yargs.version(false);
   }
 
@@ -106,6 +113,56 @@ export default class Deploy extends YargsCommand {
     }
 
     CliTelemetry.sendTelemetryEvent(TelemetryEvent.Deploy, {
+      [TelemetryProperty.Success]: TelemetrySuccess.Yes,
+      ...makeEnvRelatedProperty(rootFolder, inputs),
+    });
+    return ok(null);
+  }
+}
+
+class DeployManifest extends YargsCommand {
+  public readonly commandHead = "manifest";
+  public readonly command = this.commandHead;
+  public readonly description = "Synchronize manifest changes to Teams.";
+
+  public builder(yargs: Argv): Argv<any> {
+    this.params = HelpParamGenerator.getYargsParamForHelp("update");
+    return yargs.version(false).options(this.params);
+  }
+
+  public async runCommand(args: {
+    [argName: string]: string | string[];
+  }): Promise<Result<null, FxError>> {
+    const rootFolder = path.resolve((args.folder as string) || "./");
+    CliTelemetry.withRootFolder(rootFolder).sendTelemetryEvent(TelemetryEvent.UpdateManifestStart);
+
+    const result = await activate(rootFolder);
+    if (result.isErr()) {
+      CliTelemetry.sendTelemetryErrorEvent(TelemetryEvent.UpdateManifest, result.error);
+      return err(result.error);
+    }
+    const core = result.value;
+    let inputs: Inputs;
+    {
+      const func: Func = {
+        namespace: "fx-solution-azure/fx-resource-appstudio",
+        method: "updateManifest",
+      };
+
+      inputs = getSystemInputs(rootFolder, args.env as any);
+      const result = await core.executeUserTask!(func, inputs);
+      if (result.isErr()) {
+        CliTelemetry.sendTelemetryErrorEvent(
+          TelemetryEvent.UpdateManifest,
+          result.error,
+          makeEnvRelatedProperty(rootFolder, inputs)
+        );
+
+        return err(result.error);
+      }
+    }
+
+    CliTelemetry.sendTelemetryEvent(TelemetryEvent.UpdateManifest, {
       [TelemetryProperty.Success]: TelemetrySuccess.Yes,
       ...makeEnvRelatedProperty(rootFolder, inputs),
     });
