+++ packages/cli/src/cmds/index.ts
@@ -16,7 +16,6 @@ import Publish from "./publish";
 import Package from "./package";
 import Config from "./config";
 import Preview from "./preview/preview";
-import Manifest from "./manifest";
 import { isRemoteCollaborationEnabled } from "../utils";
 import Permission from "./permission";
 import Env from "./env";
@@ -32,7 +31,6 @@ export const commands: YargsCommand[] = [
   new Provision(),
   new Deploy(),
   new Package(),
-  new Manifest(),
   new ManifestValidate(),
   new Publish(),
   new Config(),
