--- packages/cli/src/cmds/manifest.ts
@@ -1,93 +1,0 @@
-// Copyright (c) Microsoft Corporation.
-// Licensed under the MIT license.
-
-"use strict";
-
-import * as path from "path";
-import { Argv } from "yargs";
-import { FxError, err, ok, Result, Func, Inputs } from "@microsoft/teamsfx-api";
-import activate from "../activate";
-import { YargsCommand } from "../yargsCommand";
-import { getSystemInputs, toLocaleLowerCase } from "../utils";
-import CliTelemetry, { makeEnvRelatedProperty } from "../telemetry/cliTelemetry";
-import {
-  TelemetryEvent,
-  TelemetryProperty,
-  TelemetrySuccess,
-} from "../telemetry/cliTelemetryEvents";
-import HelpParamGenerator from "../helpParamGenerator";
-
-export default class Manifest extends YargsCommand {
-  public readonly commandHead = "manifest";
-  public readonly command = `${this.commandHead} <action>`;
-  public readonly description =
-    "Manage Teams app manifest. The supported actions are 'validate' and 'update'.";
-
-  public readonly subCommands: YargsCommand[] = [new ManifestUpdate()];
-
-  public builder(yargs: Argv): Argv<any> {
-    yargs.options("action", {
-      description: `${this.subCommands.map((cmd) => cmd.commandHead).join("|")}`,
-      type: "string",
-      choices: this.subCommands.map((cmd) => cmd.commandHead),
-    });
-    this.subCommands.forEach((cmd) => {
-      yargs.command(cmd.command, cmd.description, cmd.builder.bind(cmd), cmd.handler.bind(cmd));
-    });
-    return yargs.version(false);
-  }
-
-  public async runCommand(args: { [argName: string]: string }): Promise<Result<null, FxError>> {
-    return ok(null);
-  }
-}
-
-class ManifestUpdate extends YargsCommand {
-  public readonly commandHead = "update";
-  public readonly command = this.commandHead;
-  public readonly description = "Synchronize manifest changes to Teams.";
-
-  public builder(yargs: Argv): Argv<any> {
-    this.params = HelpParamGenerator.getYargsParamForHelp("update");
-    return yargs.version(false).options(this.params);
-  }
-
-  public async runCommand(args: {
-    [argName: string]: string | string[];
-  }): Promise<Result<null, FxError>> {
-    const rootFolder = path.resolve((args.folder as string) || "./");
-    CliTelemetry.withRootFolder(rootFolder).sendTelemetryEvent(TelemetryEvent.UpdateManifestStart);
-
-    const result = await activate(rootFolder);
-    if (result.isErr()) {
-      CliTelemetry.sendTelemetryErrorEvent(TelemetryEvent.UpdateManifest, result.error);
-      return err(result.error);
-    }
-    const core = result.value;
-    let inputs: Inputs;
-    {
-      const func: Func = {
-        namespace: "fx-solution-azure/fx-resource-appstudio",
-        method: "updateManifest",
-      };
-
-      inputs = getSystemInputs(rootFolder, args.env as any);
-      const result = await core.executeUserTask!(func, inputs);
-      if (result.isErr()) {
-        CliTelemetry.sendTelemetryErrorEvent(
-          TelemetryEvent.UpdateManifest,
-          result.error,
-          makeEnvRelatedProperty(rootFolder, inputs)
-        );
-
-        return err(result.error);
-      }
-    }
-
-    CliTelemetry.sendTelemetryEvent(TelemetryEvent.UpdateManifest, {
-      [TelemetryProperty.Success]: TelemetrySuccess.Yes,
-      ...makeEnvRelatedProperty(rootFolder, inputs),
-    });
-    return ok(null);
-  }
-}
