+++ packages/cli/tests/unit/cmds/index.tests.ts
@@ -38,7 +38,6 @@ describe("Register Commands Tests", function () {
     expect(registeredCommands).includes("provision");
     expect(registeredCommands).includes("deploy");
     expect(registeredCommands).includes("package");
-    expect(registeredCommands).includes("manifest");
     expect(registeredCommands).includes("validate");
     expect(registeredCommands).includes("publish");
     expect(registeredCommands).includes("config");
